[snomed] Fix interference of object handler with invalidation events
The registered CDOObjectHandler reloaded the backing revision of an
object, which was about to change into PROXY state. This turned the
object's state back to CLEAN, but the reloaded revision was set to null
in the following parts of the invalidation method. This resulted in an
IllegalStateException being thrown when the object was accessed later.

Fixes SO-1619.