SO-1652: fixed timeout issue in event API
In very rare circumstances the promise returned by the API is already
resolved (due to faster execution in the event handler thread than the
origin thread), and the execution of callbacks - added after sending the
message - were skipped, resulting in TimeoutException on the origin
thread.
Refactored the event API and moved the p2p communication to the Event
hierarchy in the form of send(IEventBus, Class) method.
Also, using Guava's Futures utility class to add listener to the
promise/future instead of reinventing the wheel.