[core] Use IStoreAccessor from the session for change processors
The "canCopyThreadLocal" mechanism is deprecated, as we are no longer
running change processors outside a commit. In such cases,
StoreThreadLocal is set to the server session of the committer for the
job that runs the change processor in DelegateCDOServerChangeManager.

If the change processor also uses multiple threads, those threads should
set the thread-local IStoreAccessor associated with this session and
make sure that it is released when they are done. 

The session should only be released by DelegateCDOServerChangeManager.