[core] lock parent branch when creating/reopening branches
This will ensure that we create only a single child branch at any point
in time under a given parent (creating a branch takes less than a
second, this won't cause any performance problems).

Reason: 
Before v5.0 (flat index) we allowed only a single thread to initialize
the required data structures in a revision index, which means we only
allowed a single branch to be created under a given parent branch. Since
we moved this kind of data structures to the branch document in flat
index, we must ensure the same behavior when (re)opening branches to
prevent race conditions.

See related v4.6.4 code with synchronized block:
https://github.com/b2ihealthcare/snow-owl/blob/v4.6.4/core/com.b2international.snowowl.datastore.server/src/com/b2international/snowowl/datastore/server/index/IndexServerService.java#L226

Example scenario:
1. Thread A would like to create a new child ('a') under branch MAIN
2. Thread B would like to create a new child ('b') under branch MAIN
3. Both will read the latest segment info from MAIN and use it. This
will lead to invalid parent segment identifiers on at least one of the
child branches (one segmentId will be missing from the parentSegments
array).
4. On the other hand the active segment assigned to the parent might not
be one that should be (it will either be segmentId generated by the
creation of 'a' or 'b', but it's not deterministic).