TCC-15 fixing refset export file names
when a concept's description are preferred in multiple languages, an
explicit first ordering is created to prioritize the desired languages.
Explicit first meaning that the desired language codes are starting the
list, and the remaining, not mentioned language codes are following
them. So to pick a desired acceptability member, we need to look for a
min() in such list of language code, instead of max.