APDS-131: Handle back and forth donation
- added tests to handle use-cases when an extension takes an already
released international component, makes changes on it and then the
changes are "propagated" back to the international release.
- for allowing such cases to go through the upgrade process effective
time conflict handling had to be changed. From now on if there is an
effective time change conflict during any merge/rebase the component
with later effective time will be accepted and the older version and all
of it's changes will be ignored.


https://jira.ihtsdotools.org/browse/APDS-131

Conflicts:
	core/com.b2international.snowowl.datastore.server/src/com/b2international/snowowl/datastore/server/cdo/AbstractCDOConflictProcessor.java
	core/com.b2international.snowowl.datastore.server/src/com/b2international/snowowl/datastore/server/cdo/ICDOConflictProcessor.java
	core/com.b2international.snowowl.datastore.server/src/com/b2international/snowowl/datastore/server/internal/branch/CDOBranchMerger.java