[core] allow reuse of deleted branches via branch create API
Using the same name and parent you can reopen a deleted branch by
sending a create request to the repository. This will open a new
CDOBranch and override the currently stored deleted branch document in
the index. The old CDOBranch will remain in the DB but won't be
accessible anymore via the API.