SO-3044: improve index level branching services
Move much of the existing functionality to index.api module.
Move basic test cases to index.api.tests.
Refactor how branch segments are calculated and maintained. Instead of
keeping track of segment IDs, we use the base and head timestamps of
branches to keep track of segments (see RevisionSegment and
RevisionBranchPoint). On Revision documents instead of keeping track of
segmentId (essentially the segment that created the revision) and a set
of replacedIn segment IDs (segments that marked this revision deleted),
we keep track of segment ranges in IPv6 Strings (essentially 128 bit,
2x64 bit long values, otherwise query performance would drop
significantly).
Migrate much of the services to use the new segment structure (purge
functionality is still WIP).