[reasoner] Multiple changes to DelegateOntology
- Don't generate OWL entity declaration axioms. The type of each OWL
entity is gathered from other OWL axioms it participates in; punning
also requires to not tie entities to a certain type.

In addition to the above, the functional OWL export format will include
such declarations automatically, so it is also redundant to generate
them here.

- Remove OWL data property-based representation of concrete domain
members. This functionality was already commented out in favor of the
pre-existing "value as atomic concept" approach, which is more
performant and also works with reasoners that do not support "has data"
class expressions.

- Add initial support for exporting SNOMED CT terms along with
"structural" content in OWL files, as rdfs:label annotations (no
language code/dialect mapping; last visited active FSN is exported for
each concept)