[index] fix some of the type_removal warnings reported by ES
The remaining warnings thrown by the Mapping API, but we cannot remove
it yet, due to our dependency to the TCP client infrastructure. Those
warnings will be eliminated once ES 8 is out and we decide to upgrade to
that (which will remove the TCP client functionality entirely).
Fix some java compile warnings as well.
And also add some serialVersionUIDs.