[api] introduce CodeSystemURI and CodeSystemResourceRequest
Support `build(uri)` request builder method on selected SNOMED CT APIs.
This allows clients to call the API with a CodeSystemURI String
representation instead of specifying the repositoryId and branch to
delegate the request to.
CodeSystemURI represents a String in format of `<CODESYSTEM>[/<PATH>]`,
where:
- `<CODESYSTEM>` is the shortName ID of a Code System
- `<PATH>` is one of the following values:
  * `LATEST` - special value that represents the latest released version
of the codesystem. This is the default value if PATH is omitted.
Examples: `SNOMEDCT` (implicit latest) or `SNOMEDCT/LATEST` (explicit
latest).
  * `HEAD` - special value that represents the latest development
version of the codesystem. Examples: `SNOMEDCT/HEAD`
  * `<versionId>` - an explicit `versionId` that matches one existing
version of the Code System. Examples: `SNOMEDCT/2019-01-31` or
`SNOMEDCT-UK/2019-10-31`
  * `<branch_path>` - any other path value will be treated as relative
path to the CodeSystem's current working branch
(`CodeSystem.branchPath`). Examples: `SNOMEDCT/a/b`.

This new feature is available in all endpoints that require a `/:path`
parameter.