SO-3948: convert jobId to jobKey
Supplied id is no longer acts as a unique key, but as a basis to
generate the unique SHA-1 key of a remote job.
This will ensure that the unique key of the job is URL safe and can be
returned safely from REST API endpoints.
Add filterByKey/filterByKeys method to SearchJobRequestBuilder.
Support 'restart' option in ScheduleJobRequests. With this option the
caller can restart (replace in fact) an already
completed/failed/canceled job with the same key easily (will eliminate a
lot of boilerplate code around clients).