[index] fix branch synchronization issue
Properly calculate the state of a branch compared to another branch even
if the branch has been synchronized with the other branch a couple of
times. Take actual commits into account before deciding whether a branch
is in forward or behind state.
After applying changes to the index writer double check whether it has
any actual changes in it, if not, do not create an empty commit object
Empty, non-merge commits incorrectly modify the branch state, which
require unnecessary compare and merge operations between the branches.