feat(auth): support oath2 style bearer token headers in rest tests
Support security filter when searching for resource documents
(CodeSystems, Bundles, etc.) which restricts the result set to the
allowed resources.
Users require either direct access by ID, or transitive access via
permission to a container bundle ID otherwise they will receive HTTP 404
messages that the resource is not found (or empty lists if it was a
search request).