feat(core): Re-enable searchAfter in ConceptSuggestionRequest
Only a single resource is queried and so this request parameter should
be valid on the input side, also when returned with search results.

On the other hand, ConceptSearchRequest with returning results from
multiple code systems may not use this parameter and will receive a
`null` value in the result set.