feat(snomed): introduce NamespaceIdProvider to extract namespace values.
...from a collection of namespaceConceptIds

Introduce namespaceConceptId in SNOMED CT component create endpoints to
support creation of SNOMED CT content by specifying the
namespaceConceptId as opposed to the actual namespace value. The system
will be able to extract this information from the namespace concept and
use it to generate the new IDs with the correct namespace.