refactor(commons): Rename project to avoid M2e test project detection
In newer releases of M2e, when synchronizing Maven projects with their
POM that are also Eclipse plug-ins, source folders are converted to test
sources if the project name includes "tests" anywhere in the string, or
ends with "test" or "tests".

According to this rule, c.b.s.test.commons will not have test sources,
but the source folder of c.b.index.tests.tools will be converted to
test-only on synchronization; since the former bundle depends on the
latter, a build path error of "Project has only main sources but depends
on project ... which has only test sources" appears.

Changing the project name allows us to keep the source folder a regular
one without having to fight the tooling constantly.