refactor(core): Clarify that resources store bundle ancestor IDs...
...in order of depth, ie. a bundle path/absolute path can be computed by
concatenating all ancestor IDs, the container bundle ID and optionally,
the resource ID.

The only exception are resources that reside in the resource root, where
both the container bundle ID is set to -1 and the ancestor ID array
contains a single element with -1 as the value.

Additionally, the descendant updating step in BaseResourceUpdateRequest
was refactored to process siblings in a single iteration, and reduce
memory usage slightly.