feat(core): Restore tracking of lock targets held by a client
Most of the operations lock a branch or repository target only for a
short time interval eg. during a commit operation, but for some
authoring tasks we allow clients to issue lock requests remotely. These
locks should be released if the client disconnects and does not clean up
on its own.