fix(commons): Wrap Groovy 3.0.9 and extensions into an OSGi bundle
While all JARs include an OSGi-compatible manifest, extensions such as
XML, JSON and text template helper classes were converted to bundle
fragments by the authors.

This causes problems when we want to refer to eg. an XML builder class
in a script, as we can only require the "core" Groovy bundle to be
present, we can't ask for a fragment to be added to the classpath.

(Importing the Java package would usually help, however in this case the
same package exists both in the core bundle and in the fragment. So this
also complicates matters.)