feat(api): add batchSize configuration parameter to RF2 import API
Allow clients to configure the number of concepts to be indexed in each
low level Elasticsearch commit. This number is approximate and it is
used to configure the maximum size of a strongly connected component
graph in each effective time slice.
Previously this was hardcoded to 60_000 concepts, so the default value
is 60_000.