feat(core): introduce branch snapshot and realtime content requests
This was previously supported by the RevisionIndexReadRequest class but
instead of supporting it in one class with a boolean switch, now we have
two separate specific classes.
Removing RepositoryContext.openBranch functionality and replacing it
with equivalent low-level API calls and checks to eliminate the extra
branch fetches when checking availability of an accessed branch.