fix(authz): ensure we are always checking the logical ID first then...
...checking full URIs

Also make sure we are checking the base resource URI first for access
and then checking the version/branch access if needed. Most of the time
clients will have direct access to the entire resource and not just to
the version. This eliminates an unnecessary permission check call to the
authorization service regarding the accessed versioned URI but we
already have a access to the resource itself.